Install RSync before running the deploy script

+++ scripts/deploy.sh
@@ -1,5 +1,10 @@
 #!/usr/bin/env bash
 
+if ! [ -x "$(command -v rsync)" ]; then
+  echo 'Error: rsync is not installed.' >&2
+  exit 1
+fi
+
 npm run build
 ssh -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null $DEPLOY_USER@$DEPLOY_HOST -p$DEPLOY_SSH_PORT 'rm -rf dash.gonevis.com/*'
 rsync -avz -e "ssh -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null -p$DEPLOY_SSH_PORT" --progress .htaccess dist/* $DEPLOY_USER@$DEPLOY_HOST:$DEPLOY_PATH
